+++ work1/ksx919/AnimalNotFoundException.java
@@ -1,0 +1,5 @@
+public class AnimalNotFoundException extends RuntimeException {
+    public AnimalNotFoundException(String message) {
+        super(message);
+    }
+}
