+++ work1/ksx919/AnimalShop.java
@@ -1,0 +1,5 @@
+public interface AnimalShop{
+    void buyAnimal(Animal animal);
+    void serveCustomer(Customer customer,Animal animal) throws AnimalNotFoundException;
+    void closeShop();
+}
