+++ work1/ksx919/InsufficientBalanceException.java
@@ -1,0 +1,5 @@
+public class InsufficientBalanceException extends RuntimeException {
+    public InsufficientBalanceException(String message) {
+        super(message);
+    }
+}
